# social-flow-backend/app/api/v1/endpoints/analytics.py
"""
Analytics endpoints.

This module contains all analytics-related API endpoints.
"""

from typing import Any, Dict, List, Optional
from fastapi import APIRouter, Depends, HTTPException, Query
from sqlalchemy.ext.asyncio import AsyncSession

from app.core.database import get_db
from app.core.exceptions import AnalyticsServiceError
from app.models.user import User
from app.api.v1.endpoints.auth import get_current_active_user
from app.services.analytics_service import analytics_service

router = APIRouter()


@router.post("/track")
async def track_event(
    event_type: str,
    event_data: dict,
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Track an analytics event."""
    try:
        # Add user context to event data
        event_data["user_id"] = str(current_user.id)
        event_data["timestamp"] = datetime.utcnow().isoformat()
        
        # Track the event
        await analytics_service.track_event(
            event_type=event_type,
            user_id=str(current_user.id),
            data=event_data
        )
        
        return {
            "message": "Event tracked successfully",
            "event_type": event_type,
            "timestamp": event_data["timestamp"]
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to track event")


@router.get("/")
async def get_analytics(
    time_range: str = Query("7d"),
    metric_type: str = Query("overview"),
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Get analytics data."""
    try:
        result = await analytics_service.get_analytics(
            user_id=str(current_user.id),
            time_range=time_range,
            metric_type=metric_type
        )
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to get analytics")


@router.get("/user/{user_id}")
async def get_user_analytics(
    user_id: str,
    time_range: str = Query("7d"),
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Get user-specific analytics."""
    try:
        result = await analytics_service.get_user_analytics(
            user_id=user_id,
            time_range=time_range
        )
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to get user analytics")


@router.get("/content/{content_id}")
async def get_content_analytics(
    content_id: str,
    time_range: str = Query("7d"),
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Get content-specific analytics."""
    try:
        result = await analytics_service.get_content_analytics(
            content_id=content_id,
            time_range=time_range
        )
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to get content analytics")


@router.get("/reports/{report_type}")
async def get_report(
    report_type: str,
    time_range: str = Query("7d"),
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Get analytics report."""
    try:
        result = await analytics_service.generate_report(
            report_type=report_type,
            parameters={"time_range": time_range}
        )
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to generate report")


# Enhanced analytics endpoints from Scala service

@router.post("/streaming/process")
async def process_streaming_analytics(
    event_data: dict,
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Process real-time streaming analytics data."""
    try:
        result = await analytics_service.process_streaming_analytics(event_data)
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to process streaming analytics")


@router.get("/realtime/{metric_type}")
async def get_real_time_metrics(
    metric_type: str,
    time_window: str = Query("1m"),
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Get real-time metrics for monitoring."""
    try:
        result = await analytics_service.get_real_time_metrics(
            metric_type=metric_type,
            time_window=time_window
        )
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to get real-time metrics")


@router.get("/business-intelligence/{report_type}")
async def get_business_intelligence(
    report_type: str,
    time_range: str = Query("30d"),
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Get business intelligence reports."""
    try:
        result = await analytics_service.get_business_intelligence(
            report_type=report_type,
            time_range=time_range
        )
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to get business intelligence")


@router.get("/anomaly-detection/{metric_type}")
async def get_anomaly_detection(
    metric_type: str,
    threshold: float = Query(2.0),
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Detect anomalies in metrics."""
    try:
        result = await analytics_service.get_anomaly_detection(
            metric_type=metric_type,
            threshold=threshold
        )
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to detect anomalies")


@router.get("/predictive/{prediction_type}")
async def get_predictive_analytics(
    prediction_type: str,
    horizon: str = Query("7d"),
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Get predictive analytics forecasts."""
    try:
        result = await analytics_service.get_predictive_analytics(
            prediction_type=prediction_type,
            horizon=horizon
        )
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to get predictive analytics")


@router.get("/dashboard")
async def get_dashboard_data(
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Get dashboard analytics data."""
    try:
        # Get multiple analytics data for dashboard
        overview = await analytics_service.get_analytics(
            user_id=str(current_user.id),
            time_range="7d",
            metric_type="overview"
        )
        
        realtime = await analytics_service.get_real_time_metrics(
            metric_type="user_activity",
            time_window="1h"
        )
        
        return {
            "overview": overview,
            "realtime": realtime,
            "timestamp": datetime.utcnow().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to get dashboard data")


@router.get("/export/{report_type}")
async def export_analytics(
    report_type: str,
    format: str = Query("json"),
    time_range: str = Query("30d"),
    current_user: User = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """Export analytics data."""
    try:
        # TODO: Implement analytics data export
        # This would generate downloadable reports in various formats
        
        return {
            "message": "Export functionality not yet implemented",
            "report_type": report_type,
            "format": format,
            "time_range": time_range
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail="Failed to export analytics")