# social-flow-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Social Flow Backend API
  description: |
    A comprehensive social media backend API combining YouTube and Twitter features
    with advanced AI/ML capabilities. This API supports video streaming, social posts,
    real-time interactions, content moderation, analytics, and machine learning features.
  version: 1.0.0
  contact:
    name: Social Flow Team
    email: support@socialflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.socialflow.com/api/v1
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /analytics/track:
    post:
      tags:
        - analytics
      summary: Track analytics event
      description: Track user analytics event
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_type:
                  type: string
                  description: Type of event to track
                data:
                  type: object
                  description: Event data
              required:
                - event_type
      responses:
        '200':
          description: Event tracked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success, failed]
                  message:
                    type: string
  /analytics/user/{user_id}:
    get:
      tags:
        - analytics
      summary: Get user analytics
      description: Get analytics data for a specific user
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: time_period
          in: query
          schema:
            type: string
            enum: [1d, 7d, 30d]
            default: 7d
      responses:
        '200':
          description: User analytics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  time_period:
                    type: string
                  metrics:
                    type: object
                  events_count:
                    type: integer
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...